Add symlink and hardlink support for file copy
Data files (i.e. non-processed, non-Markdown files) are usually copied
twice, possibly thrice: from the vault to a temporary dir (if enabled),
from that dir to the Markdown dir, then finally to the HTML dir. This is
incredibly wasteful and dramatically slows down builds with lots of
ancillary data files. Linking instead of copying saves on storage space
and speed.